requirements.txt
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==44.0.3
defusedxml==0.7.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.10
flake8==7.2.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
mccabe==0.7.0
mypy_extensions==1.1.0
oauthlib==3.2.2
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
python-decouple==3.8
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
